requirements.txt
# Discord RSVP Bot - Python Dependencies

# Discord API library
disnake>=2.9.0

# Supabase Python client
supabase>=2.3.0

# Environment variable management
python-dotenv>=1.0.0

# Async file operations (for backup system)
aiofiles>=23.0.0

# HTTP client for external API calls
aiohttp>=3.8.0

# Timezone handling
pytz>=2023.3

# System monitoring (for bot monitoring features)
psutil>=5.9.0

# Data validation and serialization
pydantic>=2.0.0

# Enhanced logging and monitoring
structlog>=23.0.0

# Development dependencies (optional - uncomment if needed)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0 